Add a capability, settings, and improve a behavior.
Add hidden hash tags fields and feature. User can now add hashtags to be
appended to the tweet, but without typing them in manually. They're
hidden hash tags.  Add in requisite settings and fields to make this
work.

Improve behavior of removal checkboxes in shortcode creation dialog.
They now actually remove the element from the tweet the user sends, not
just the preview in the dialog.  This is what they SHOULD have done
initially, but didn't.  It was confusing and largely pointless before.