Add tweet templating per feature request.
This is more complicated than it sounds... It obviously included adding
template settings, a template parsing engine, and related code. However,
it also led me to generate the shortcode creator dialog tweet preview
using the code that actually generates the tweet (as opposed to a
separate JavaScript version solely for the preview) which required AJAX.
In the long run, this makes maintainability and extendability easier
(and is thus better), but it meant lots of work and there is now a delay
for the preview.